In order to change the code structure of the project.
1. Move the java code to a separate folder at the root
2. Update the GitHub MD pages under the docs folder
