+++ .gitignore
@@ -27,7 +27,7 @@
 hs_err_pid*
 
 # Maven output dir
-target
+*/target/
 
 
 ###############################################################
