+++ docs/1_Introduction.md
@@ -34,7 +34,7 @@ and physical books.
 
 Open the code in your favorite IDE:
 
-* To work on the java version, simply load the root [pom.xml](../pom.xml) in 
+* To work on the java version, simply load the root [pom.xml](../java/pom.xml) in 
   your favourite IDE
 * We currently only have a Visual Studio environment around the C++ code (BTW,
   contributions are welcome :-)). Open [testbuilders-workshop.sln](../testbuilders-workshop.sln).
