+++ docs/4_Mikado_Method_to_the_Rescue.md
@@ -188,7 +188,7 @@ So let's delete the tests!
 ### Use Mikado Method to create Test Data Builders to add a test on Invoice
 
 We'll first focus on the test on
-[Invoice](../src/main/java/com/murex/tbw/purchase/Invoice.java). It's simpler
+[Invoice](../java/src/main/java/com/murex/tbw/purchase/Invoice.java). It's simpler
 and achievable given the time we have.
 
 Write your test as if you had the builders, and start to draw your mikado
@@ -200,7 +200,7 @@ something useful.
 ### Use Mikado Method to create Test Data Builders to add a test on Report Generator (advanced)
 
 If you have the time, repeat the same exercise with the test on 
-[ReportGenerator](../src/main/java/com/murex/tbw/report/ReportGenerator.java)
+[ReportGenerator](../java/src/main/java/com/murex/tbw/report/ReportGenerator.java)
 
 You should be able to re-use many builders you wrote before.
