+++ java/src/main/java/com/murex/tbw/domain/book/Author.java
@@ -1,0 +1,56 @@
+/*******************************************************************************
+ *
+ * Copyright (c) {2003-2019} Murex S.A.S. and its affiliates.
+ * All rights reserved. This program and the accompanying materials
+ * are made available under the terms of the MIT License
+ * which accompanies this distribution, and is available at
+ * https://opensource.org/licenses/MIT
+ *
+ *******************************************************************************/
+
+package com.murex.tbw.domain.book;
+
+import com.murex.tbw.domain.country.Country;
+
+import java.util.Objects;
+
+public final class Author {
+
+    private final String name;
+    private final Country nationality;
+
+    public Author(String name, Country nationality) {
+        this.name = name;
+        this.nationality = nationality;
+    }
+
+    public String getName() {
+        return name;
+    }
+
+    public Country getNationality() {
+        return nationality;
+    }
+
+    @Override
+    public String toString() {
+        return "Author{" +
+                "name='" + name + '\'' +
+                ", nationality=" + nationality +
+                '}';
+    }
+
+    @Override
+    public boolean equals(Object o) {
+        if (this == o) return true;
+        if (o == null || getClass() != o.getClass()) return false;
+        Author author = (Author) o;
+        return Objects.equals(name, author.name) &&
+                Objects.equals(nationality, author.nationality);
+    }
+
+    @Override
+    public int hashCode() {
+        return Objects.hash(name, nationality);
+    }
+}
