+++ java/src/main/java/com/murex/tbw/domain/country/Currency.java
@@ -1,0 +1,20 @@
+/*******************************************************************************
+ *
+ * Copyright (c) {2003-2019} Murex S.A.S. and its affiliates.
+ * All rights reserved. This program and the accompanying materials
+ * are made available under the terms of the MIT License
+ * which accompanies this distribution, and is available at
+ * https://opensource.org/licenses/MIT
+ *
+ *******************************************************************************/
+
+package com.murex.tbw.domain.country;
+
+public enum Currency {
+    US_DOLLAR,
+    AUSTRALIAN_DOLLAR,
+    EURO,
+    POUND_STERLING,
+    YEN,
+    RENMINBI
+}
